requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
email-validator
python-jose[cryptography]
passlib
python-multipart
